requirements.txt
python-telegram-bot>=20.0
pyppeteer>=1.0.2
jinja2>=3.0.0
python-dotenv>=0.19.0
pillow>=8.0.0
reportlab>=3.6.0
